requirements.txt
fastapi
uvicorn
pydantic
websocket-client
pillow
requests
python-multipart